requirements.txt
slack_sdk>=3.0.0
python-dotenv>=0.19.0
